--- src/main/java/com/errday/springsecuritystudy/annotation/OwnerShip.java
@@ -1,19 +1,0 @@
-package com.errday.springsecuritystudy.annotation;
-
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PreAuthorize;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Documented
-@Retention(RUNTIME)
-@Target({TYPE, METHOD})
-@PostAuthorize("returnObject.owner == authentication.name")
-public @interface OwnerShip {
-}
