--- src/main/java/com/errday/springsecuritystudy/component/CustomAuthorizer.java
@@ -1,12 +1,0 @@
-package com.errday.springsecuritystudy.component;
-
-import org.springframework.security.access.expression.method.MethodSecurityExpressionOperations;
-import org.springframework.stereotype.Component;
-
-@Component("myAuthorizer")
-public class CustomAuthorizer {
-
-    public boolean isUser(MethodSecurityExpressionOperations root) {
-        return root.hasAuthority("ROLE_USER");
-    }
-}
