--- src/main/java/com/errday/springsecuritystudy/config/AnonymousConfig.java
@@ -1,33 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-//@Configuration
-//@EnableWebSecurity
-public class AnonymousConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers("/anonymous").hasRole("GUEST")
-                .requestMatchers("/anonymousContext", "/authentication").permitAll()
-                .anyRequest().authenticated()
-            )
-            .formLogin(Customizer.withDefaults())
-            .anonymous(anonymous -> anonymous
-                .principal("guest")
-                .authorities("ROLE_GUEST")
-            );
-        return http.build();
-    }
-
-}
