--- src/main/java/com/errday/springsecuritystudy/config/AuthenticationProviderConfig3.java
@@ -1,49 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import com.errday.springsecuritystudy.provider.CustomDaoAuthenticationProvider;
-import com.errday.springsecuritystudy.provider.CustomDaoAuthenticationProvider2;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-//@Configuration
-//@EnableWebSecurity
-public class AuthenticationProviderConfig3 {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        AuthenticationManagerBuilder managerBuilder = http.getSharedObject(AuthenticationManagerBuilder.class);
-        managerBuilder.authenticationProvider(customAuthenticationProvider());
-        managerBuilder.authenticationProvider(customAuthenticationProvider2());
-
-
-        http.authorizeHttpRequests(auth -> auth
-                .anyRequest().authenticated())
-                .formLogin(Customizer.withDefaults());
-        return http.build();
-    }
-
-    @Bean
-    public AuthenticationProvider customAuthenticationProvider() {
-        return new CustomDaoAuthenticationProvider();
-    }
-
-    @Bean
-    public AuthenticationProvider customAuthenticationProvider2() {
-        return new CustomDaoAuthenticationProvider2();
-    }
-
-}
