--- src/main/java/com/errday/springsecuritystudy/config/AuthorizeHttpRequestsConfig.java
@@ -1,41 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.servlet.util.matcher.MvcRequestMatcher;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.security.web.util.matcher.RegexRequestMatcher;
-import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
-
-//@Configuration
-//@EnableWebSecurity
-public class AuthorizeHttpRequestsConfig {
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http, HandlerMappingIntrospector introspector) throws Exception {
-
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers("/").permitAll()
-                .requestMatchers("/user").hasAuthority("ROLE_USER")
-                .requestMatchers("/myPage/**").hasRole("USER")
-                .requestMatchers(HttpMethod.POST).hasAuthority("ROLE_WRITE")
-                .requestMatchers(new AntPathRequestMatcher("/manager/**")).hasAuthority("ROLE_MANAGER")
-                .requestMatchers(new MvcRequestMatcher(introspector, "/admin/payment")).hasAuthority("ROLE_ADMIN")
-                .requestMatchers("/admin/**").hasAnyAuthority("ROLE_ADMIN", "ROLE_MANAGER")
-                .requestMatchers(new RegexRequestMatcher("/resource/[A-Za-z0-9]+", null)).hasAuthority("ROLE_MANAGER")
-                .anyRequest().authenticated()
-            )
-            .formLogin(Customizer.withDefaults())
-            .csrf(AbstractHttpConfigurer::disable)
-        ;
-
-        return http.build();
-    }
-
-}
