--- src/main/java/com/errday/springsecuritystudy/config/CustomExpressionConfig.java
@@ -1,66 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import com.errday.springsecuritystudy.component.CustomRequestMatcher;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.access.expression.DefaultHttpSecurityExpressionHandler;
-import org.springframework.security.web.access.expression.WebExpressionAuthorizationManager;
-
-//@Configuration
-//@EnableWebSecurity
-public class CustomExpressionConfig {
-
-    /*@Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers("/user/{name}").access(new WebExpressionAuthorizationManager("#name == authentication.name"))
-                .requestMatchers("/admin/db").access(new WebExpressionAuthorizationManager("hasAuthority('ROLE_DB') or hasAuthority('ROLE_ADMIN')"))
-                .anyRequest().authenticated())
-            .formLogin(Customizer.withDefaults())
-            .csrf(AbstractHttpConfigurer::disable)
-        ;
-
-        return http.build();
-    }*/
-
-    /*@Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http, ApplicationContext context) throws Exception {
-
-        DefaultHttpSecurityExpressionHandler expressionHandler = new DefaultHttpSecurityExpressionHandler();
-        expressionHandler.setApplicationContext(context);
-
-        WebExpressionAuthorizationManager authorizationManager
-                = new WebExpressionAuthorizationManager("@customWebSecurity.check(authentication, request)");
-        authorizationManager.setExpressionHandler(expressionHandler);
-
-        http.authorizeHttpRequests(auth -> auth
-                    .requestMatchers("/custom/**").access(authorizationManager)
-                    .anyRequest().authenticated())
-                .formLogin(Customizer.withDefaults())
-                .csrf(AbstractHttpConfigurer::disable)
-        ;
-
-        return http.build();
-    }*/
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http, ApplicationContext context) throws Exception {
-
-        http.authorizeHttpRequests(auth -> auth
-                    .requestMatchers(new CustomRequestMatcher("/admin")).hasAuthority("ROLE_ADMIN")
-                    .anyRequest().authenticated())
-                .formLogin(Customizer.withDefaults())
-                .csrf(AbstractHttpConfigurer::disable)
-        ;
-
-        return http.build();
-    }
-
-}
