--- src/main/java/com/errday/springsecuritystudy/config/ExceptionHandlerConfig.java
@@ -1,52 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.session.SessionRegistry;
-import org.springframework.security.core.session.SessionRegistryImpl;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.security.web.SecurityFilterChain;
-
-import java.io.IOException;
-
-//@Configuration
-//@EnableWebSecurity
-public class ExceptionHandlerConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers( "/login").permitAll()
-                .requestMatchers("/admin").hasRole("ADMIN")
-                .anyRequest().authenticated())
-            .formLogin(Customizer.withDefaults())
-            .exceptionHandling(exception -> exception
-                /*.authenticationEntryPoint(new AuthenticationEntryPoint() {
-                    @Override
-                    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-                        System.out.println(authException.getMessage());
-                        response.sendRedirect("/login");
-                    }
-                })*/
-                .accessDeniedHandler((request, response, accessDeniedException) -> {
-                    System.out.println("exception: " + accessDeniedException.getMessage());
-                    response.sendRedirect("/denied");
-                })
-            )
-        ;
-
-        return http.build();
-    }
-
-}
