--- src/main/java/com/errday/springsecuritystudy/config/HttpSessionConfig.java
@@ -1,33 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.session.MapSession;
-import org.springframework.session.MapSessionRepository;
-import org.springframework.session.SessionRepository;
-import org.springframework.session.config.annotation.web.http.EnableSpringHttpSession;
-import org.springframework.session.web.http.CookieSerializer;
-import org.springframework.session.web.http.DefaultCookieSerializer;
-
-import java.util.concurrent.ConcurrentHashMap;
-
-//@Configuration
-//@EnableSpringHttpSession
-public class HttpSessionConfig {
-
-    @Bean
-    public CookieSerializer cookieSerializer() {
-        DefaultCookieSerializer defaultCookieSerializer = new DefaultCookieSerializer();
-        defaultCookieSerializer.setCookieName("SESSION");
-        defaultCookieSerializer.setUseHttpOnlyCookie(true);
-        defaultCookieSerializer.setUseSecureCookie(true);
-        defaultCookieSerializer.setSameSite("Lax");
-
-        return defaultCookieSerializer;
-    }
-
-    @Bean
-    public SessionRepository<MapSession> sessionRepository() {
-        return new MapSessionRepository(new ConcurrentHashMap<>());
-    }
-}
