--- src/main/java/com/errday/springsecuritystudy/config/MethodSecurityConfig.java
@@ -1,33 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity(securedEnabled = true, jsr250Enabled = true)
-public class MethodSecurityConfig {
-
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http, ApplicationContext context) throws Exception {
-
-        http.authorizeHttpRequests(auth -> auth
-                    .anyRequest().authenticated())
-                .formLogin(Customizer.withDefaults())
-                .csrf(AbstractHttpConfigurer::disable)
-        ;
-
-        return http.build();
-    }
-
-
-}
