--- src/main/java/com/errday/springsecuritystudy/config/RequestCacheConfig.java
@@ -1,45 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import org.springframework.security.web.savedrequest.SavedRequest;
-
-import java.io.IOException;
-
-//@Configuration
-//@EnableWebSecurity
-public class RequestCacheConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        HttpSessionRequestCache requestCache = new HttpSessionRequestCache();
-        requestCache.setMatchingRequestParameterName("customParam=y");
-
-        http.authorizeHttpRequests(auth -> auth.anyRequest().authenticated())
-                .formLogin(form -> form
-                        .successHandler((request, response, authentication) -> {
-                            SavedRequest savedRequest = requestCache.getRequest(request, response);
-                            String redirectUrl = savedRequest.getRedirectUrl();
-                            response.sendRedirect(redirectUrl);
-                        })
-                )
-                .requestCache(cache -> cache.requestCache(requestCache));
-        return http.build();
-    }
-
-}
