--- src/main/java/com/errday/springsecuritystudy/config/SameSiteConfig.java
@@ -1,38 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import com.errday.springsecuritystudy.filter.CsrfCookieFilter;
-import com.errday.springsecuritystudy.handler.SpaCsrfTokenRequestHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
-
-//@Configuration
-//@EnableWebSecurity
-public class SameSiteConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        http.authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/csrf", "/csrfToken", "/form", "/formCsrf" ,"/cookie", "/cookieCsrf").permitAll()
-                        .anyRequest().authenticated())
-                .formLogin(Customizer.withDefaults())
-        ;
-        return http.build();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        UserDetails user = User.withUsername("user").password("{noop}1111").roles("USER").build();
-        return new InMemoryUserDetailsManager(user);
-    }
-}
