--- src/main/java/com/errday/springsecuritystudy/config/SecurityConfig2.java
@@ -1,49 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import com.errday.springsecuritystudy.filter.CustomAuthenticationFilter;
-import com.errday.springsecuritystudy.provider.CustomAuthenticationProvider;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AnonymousAuthenticationProvider;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import java.util.List;
-
-//@Configuration
-//@EnableWebSecurity
-public class SecurityConfig2 {
-
-    /*@Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(auth -> auth
-                    .requestMatchers("/", "/api/login").permitAll()
-                    .anyRequest().authenticated())
-            .addFilterBefore(customAuthenticationFilter(http), UsernamePasswordAuthenticationFilter.class);
-        return http.build();
-    }*/
-
-   /* @Bean
-    public CustomAuthenticationFilter customAuthenticationFilter(HttpSecurity http) {
-        List<AuthenticationProvider> list1 = List.of(new DaoAuthenticationProvider());
-        ProviderManager parent = new ProviderManager(list1);
-
-        List<AuthenticationProvider> list2 = List.of(new AnonymousAuthenticationProvider("key"), new CustomAuthenticationProvider());
-        ProviderManager providerManager = new ProviderManager(list2, parent);
-
-        CustomAuthenticationFilter customAuthenticationFilter = new CustomAuthenticationFilter(http);
-        customAuthenticationFilter.setAuthenticationManager(providerManager);
-
-        return customAuthenticationFilter;
-    }*/
-
-}
