+++ src/main/java/com/errday/springsecuritystudy/config/SecuriyConfig.java
@@ -1,0 +1,41 @@
+package com.errday.springsecuritystudy.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecuriyConfig {
+
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.authorizeRequests(auth -> auth
+                .requestMatchers("/user").hasRole("USER")
+                .requestMatchers("/db").hasRole("DB")
+                .requestMatchers("/admin").hasRole("ADMIN")
+                .anyRequest().authenticated())
+            .formLogin(Customizer.withDefaults())
+            .csrf(AbstractHttpConfigurer::disable);
+        return http.build();
+    }
+
+    @Bean
+    public RoleHierarchy roleHierarchy() {
+        RoleHierarchyImpl roleHierarchy = new RoleHierarchyImpl();
+        roleHierarchy.setHierarchy(
+                "ROLE_ADMIN > ROLE_DB\n"
+                + "ROLE_DB > ROLE_MANAGER\n"
+                + "ROLE_MANAGER > ROLE_USER\n"
+                + "ROLE_USER > ROLE_ANONYMOUS\n"
+        );
+        return roleHierarchy;
+    }
+}
