--- src/main/java/com/errday/springsecuritystudy/config/SessionConfig.java
@@ -1,51 +1,0 @@
-package com.errday.springsecuritystudy.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.session.SessionRegistry;
-import org.springframework.security.core.session.SessionRegistryImpl;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-//@Configuration
-//@EnableWebSecurity
-public class SessionConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers( "/invalidSessionUrl", "/expiredUrl").permitAll()
-                .anyRequest().authenticated())
-            .formLogin(Customizer.withDefaults())
-            /*.sessionManagement(session -> session
-                //.invalidSessionUrl("/invalidSessionUrl")
-                .maximumSessions(1)
-                .maxSessionsPreventsLogin(false)
-                .expiredUrl("/expiredUrl")*/
-            /*.sessionManagement(session -> session
-                .sessionFixation(sessionFixation -> sessionFixation.changeSessionId())*/
-            /*.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.IF_REQUIRED))*/
-            .sessionManagement(session -> session
-                .maximumSessions(2)
-                .maxSessionsPreventsLogin(false)
-            )
-        ;
-
-        return http.build();
-    }
-
-    @Bean
-    public SessionRegistry sessionRegistry() {
-        return new SessionRegistryImpl();
-    }
-
-}
