--- src/main/java/com/errday/springsecuritystudy/controller/MethodController.java
@@ -1,128 +1,0 @@
-package com.errday.springsecuritystudy.controller;
-
-import com.errday.springsecuritystudy.annotation.IsAdmin;
-import com.errday.springsecuritystudy.annotation.OwnerShip;
-import com.errday.springsecuritystudy.service.DataService;
-import com.errday.springsecuritystudy.user.Account;
-import jakarta.annotation.security.DenyAll;
-import jakarta.annotation.security.PermitAll;
-import jakarta.annotation.security.RolesAllowed;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PostFilter;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-@RestController
-@RequiredArgsConstructor
-public class MethodController {
-
-    private final DataService dataService;
-
-    /*@GetMapping("/admin")
-    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
-    public String admin() {
-        return "admin";
-    }*/
-
-    /*@GetMapping("/user")
-    @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
-    public String user() {
-        return "user";
-    }*/
-
-    @GetMapping("/isAuthenticated")
-    @PreAuthorize("isAuthenticated")
-    public String isAuthenticated() {
-        return "isAuthenticated";
-    }
-
-    @GetMapping("/user/{id}")
-    @PreAuthorize("#id == authentication.name")
-    public String authentication(@PathVariable(value = "id") String id) {
-        return id;
-    }
-
-    @GetMapping("/owner")
-    @PostAuthorize("returnObject.owner == authentication.name")
-    public Account owner(String name) {
-        return new Account(name, false);
-    }
-
-    @GetMapping("/isSecure")
-    @PostAuthorize("hasAuthority('ROLE_ADMIN') and returnObject.isSecure")
-    public Account isSecure(String name, String secure) {
-        return new Account(name, "Y".equals(secure));
-    }
-
-    @PostMapping("/writeList")
-    public List<Account> writeList(@RequestBody List<Account> data) {
-        return dataService.writeList(data);
-    }
-
-    @PostMapping("/writeMap")
-    public Map<String, Account> writeMap(@RequestBody List<Account> data) {
-        Map<String, Account> accountMap = data.stream().collect(Collectors.toMap(account -> account.getOwner(), account -> account));
-        return dataService.writeMap(accountMap);
-    }
-
-    @GetMapping("/readList")
-    public List<Account> readList() {
-        return dataService.readList();
-    }
-
-    @GetMapping("/readMap")
-    public Map<String, Account> readMap() {
-        return dataService.readMap();
-    }
-
-    @GetMapping("/user")
-    @Secured("ROLE_USER")
-    public String user() {
-        return "user";
-    }
-
-    @GetMapping("/admin")
-    @RolesAllowed("ADMIN")
-    public String admin() {
-        return "admin";
-    }
-
-    @GetMapping("/permitAll")
-    @PermitAll
-    public String permitAll() {
-        return "permitAll";
-    }
-
-    @GetMapping("/denyAll")
-    @DenyAll
-    public String denyAll() {
-        return "denyAll";
-    }
-
-    @GetMapping("/isAdmin")
-    @IsAdmin
-    public String isAdmin() {
-        return "isAdmin";
-    }
-
-    @GetMapping("/ownerShip")
-    @OwnerShip
-    public Account ownerShip(String name) {
-        return new Account(name, false);
-    }
-
-
-    @GetMapping("/delete")
-    @PreAuthorize("@myAuthorizer.isUser(#root)")
-    public String delete(String name) {
-        return "delete";
-    }
-
-
-}
