--- src/main/java/com/errday/springsecuritystudy/filter/CsrfCookieFilter.java
@@ -1,25 +1,0 @@
-package com.errday.springsecuritystudy.filter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-public class CsrfCookieFilter extends OncePerRequestFilter {
-
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        CsrfToken csrfToken = (CsrfToken) request.getAttribute("_csrf");
-
-        if (csrfToken != null) {
-            csrfToken.getToken();
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
