--- src/main/java/com/errday/springsecuritystudy/filter/CustomAuthenticationFilter.java
@@ -1,46 +1,0 @@
-package com.errday.springsecuritystudy.filter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
-import org.springframework.security.web.context.DelegatingSecurityContextRepository;
-import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
-import org.springframework.security.web.context.RequestAttributeSecurityContextRepository;
-import org.springframework.security.web.context.SecurityContextRepository;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-import java.io.IOException;
-
-public class CustomAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
-
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    public CustomAuthenticationFilter(HttpSecurity http) {
-        super(new AntPathRequestMatcher("/api/login", "GET"));
-        setSecurityContextRepository(getSeucrityContextRepository(http));
-    }
-
-    private SecurityContextRepository getSeucrityContextRepository(HttpSecurity http) {
-        SecurityContextRepository securityContextRepository = http.getSharedObject(SecurityContextRepository.class);
-        if (securityContextRepository == null) {
-            securityContextRepository = new DelegatingSecurityContextRepository(
-                    new RequestAttributeSecurityContextRepository(), new HttpSessionSecurityContextRepository()
-            );
-        }
-        return securityContextRepository;
-    }
-
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException, IOException, ServletException {
-        String username = request.getParameter("username");
-        String password = request.getParameter("password");
-        UsernamePasswordAuthenticationToken token = new UsernamePasswordAuthenticationToken(username, password);
-        return getAuthenticationManager().authenticate(token);
-    }
-}
