--- src/main/java/com/errday/springsecuritystudy/handler/SpaCsrfTokenRequestHandler.java
@@ -1,30 +1,0 @@
-package com.errday.springsecuritystudy.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.security.web.csrf.CsrfTokenRequestAttributeHandler;
-import org.springframework.security.web.csrf.XorCsrfTokenRequestAttributeHandler;
-import org.springframework.util.StringUtils;
-
-import java.util.function.Supplier;
-
-public class SpaCsrfTokenRequestHandler extends CsrfTokenRequestAttributeHandler {
-
-    private CsrfTokenRequestAttributeHandler delegate = new XorCsrfTokenRequestAttributeHandler();
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, Supplier<CsrfToken> deferredCsrfToken) {
-        delegate.handle(request, response, deferredCsrfToken);
-    }
-
-    @Override
-    public String resolveCsrfTokenValue(HttpServletRequest request, CsrfToken csrfToken) {
-
-        if (StringUtils.hasText(csrfToken.getParameterName())) {
-            return super.resolveCsrfTokenValue(request, csrfToken);
-        }
-
-        return delegate.resolveCsrfTokenValue(request, csrfToken);
-    }
-}
