--- src/main/java/com/errday/springsecuritystudy/provider/CustomAuthenticationProvider.java
@@ -1,36 +1,0 @@
-package com.errday.springsecuritystudy.provider;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationProvider implements AuthenticationProvider {
-
-    private final UserDetailsService userDetailsService;
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String loginId = authentication.getName();
-        String password = authentication.getCredentials().toString();
-
-        UserDetails userDetails =  userDetailsService.loadUserByUsername(loginId);
-        if (userDetails == null) {
-            throw new UsernameNotFoundException("UsernameNotFoundException");
-        }
-
-        return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return authentication.isAssignableFrom(UsernamePasswordAuthenticationToken.class);
-    }
-}
