--- src/main/java/com/errday/springsecuritystudy/provider/CustomDaoAuthenticationProvider.java
@@ -1,28 +1,0 @@
-package com.errday.springsecuritystudy.provider;
-
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-
-import java.util.List;
-
-public class CustomDaoAuthenticationProvider implements AuthenticationProvider {
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String loginId = authentication.getName();
-        String password = authentication.getCredentials().toString();
-
-        // 아이디 검증
-        // 비밀번호 검증
-
-        return new UsernamePasswordAuthenticationToken(loginId, password, List.of(new SimpleGrantedAuthority("ROLE_USER")));
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return authentication.isAssignableFrom(UsernamePasswordAuthenticationToken.class);
-    }
-}
