--- src/main/java/com/errday/springsecuritystudy/service/DataService.java
@@ -1,39 +1,0 @@
-package com.errday.springsecuritystudy.service;
-
-import com.errday.springsecuritystudy.user.Account;
-import org.springframework.security.access.prepost.PostFilter;
-import org.springframework.security.access.prepost.PreFilter;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-@Service
-public class DataService {
-
-    @PreFilter("filterObject.owner == authentication.name")
-    public List<Account> writeList(List<Account> data) {
-        return data;
-    }
-
-    @PreFilter("filterObject.value.owner == authentication.name")
-    public Map<String, Account> writeMap(Map<String, Account> data) {
-        return data;
-    }
-
-    @PostFilter("filterObject.owner == authentication.name")
-    public List<Account> readList() {
-        return new ArrayList<>(List.of(new Account("user", false), new Account("admin", false), new Account("db", false)));
-    }
-
-    @PostFilter("filterObject.value.owner == authentication.name")
-    public Map<String, Account> readMap() {
-        return new HashMap<>(Map.of(
-                "user", new Account("user", false),
-                "admin", new Account("admin", false),
-                "db", new Account("db", false)
-        ));
-    }
-}
