--- src/main/resources/templates/cookie.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Fetch Example</title>
-    <script>
-        // fetch 요청을 실행하는 함수
-        function sendRequest() {
-            fetch('/cookieCsrf', {
-                method: 'POST', // 요청 메소드
-                headers: {
-                    'Content-Type': 'application/json', // 컨텐츠 타입 지정
-                    'X-XSRF-TOKEN': getCookie('XSRF-TOKEN') // CSRF 토큰 헤더에 추가
-                },
-                body: JSON.stringify({/* 데이터 객체 */}), // 서버로 보낼 데이터
-                credentials: 'include' // 쿠키를 포함시키기 위해 필요
-            })
-                .then(response => {
-                    if (!response.ok) {
-                        throw new Error('Network response was not ok');
-                    }
-                    return response.json(); // 여기에서 서버 응답을 처리
-                })
-                .then(data => {
-                    console.log('Success:', data); // 성공 처리 로직
-                })
-                .catch(error => {
-                    console.error('Error:', error); // 에러 처리 로직
-                });
-        }
-
-        // 쿠키 값을 가져오는 함수
-        function getCookie(name) {
-            let cookieValue = null;
-            if (document.cookie && document.cookie !== '') {
-                const cookies = document.cookie.split(';');
-                for (let i = 0; i < cookies.length; i++) {
-                    const cookie = cookies[i].trim();
-                    if (cookie.substring(0, name.length + 1) === (name + '=')) {
-                        cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                        break;
-                    }
-                }
-            }
-            return cookieValue;
-        }
-    </script>
-</head>
-<body>
-<!-- 버튼 요소 -->
-<button onclick="sendRequest()">Send Request</button>
-</body>
-</html>
