--- src/main/resources/templates/form.html
@@ -1,110 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Dashboard</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"></link>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap"></link>
-  <style>
-    body {
-      font-family: 'Roboto', sans-serif;
-      background-color: #ffffff; /* 밝은 배경색 변경 */
-      color: #000000; /* 어두운 글자색 변경 */
-    }
-    .navbar-custom {
-      background-color: #ffffff; /* 밝은 네이비 변경 */
-      color: #000000;
-      margin-bottom: 0;
-    }
-    .container-fluid {
-      padding: 0; /* container의 패딩 제거 */
-    }
-    .navbar-custom .navbar-brand, .navbar-custom .nav-link {
-      color: #000000;
-    }
-    .sidebar {
-      background-color: #ffffff; /* 밝은 네이비 변경 */
-      min-height: 100vh;
-      color: #000000;
-    }
-    .sidebar a {
-      color: #000000;
-      transition: background-color 0.3s ease, color 0.3s ease;
-    }
-    .sidebar a:hover {
-      background-color: #cccccc; /* 마우스 오버시 밝은 회색 */
-      color: #000000;
-    }
-    .content {
-      padding: 20px;
-      background-color: #ffffff; /* 밝은 그레이 변경 */
-      border-radius: 10px;
-      box-shadow: 0 0 10px rgba(0, 0, 0, 0.1); /* 박스 그림자 색상 변경 */
-      margin-top: 0;
-    }
-    .footer {
-      background-color: #ffffff; /* 밝은 네이비 변경 */
-      color: #000000;
-      padding: 10px 0;
-      border-top: 1px solid #cccccc; /* 경계선 색상 변경 */
-    }
-    .sidebar .list-group-item {
-      color: #000000; /* 글자색 변경 */
-      background-color: #ffffff; /* 배경색 변경 */
-      transition: background-color 0.3s ease, color 0.3s ease;
-    }
-
-    .sidebar .list-group-item:hover {
-      background-color: #cccccc; /* 마우스 오버시 밝은 회색 */
-      color: #000000;
-    }
-    /* 중앙 로그인 폼 스타일 변경 */
-    .login-form {
-      max-width: 400px;
-      margin: 30px auto;
-      padding: 20px;
-      background-color: #ffffff; /* 밝은 폼 배경색 변경 */
-      border-radius: 5px;
-      border: 1px solid #000000; /* 검은색 테두리 추가 */
-      box-shadow: 0 0 15px rgba(0, 0, 0, 0.1); /* 박스 그림자 색상 변경 */
-    }
-    .login-form h2 {
-      color: #000000; /* 어두운 글자색 변경 */
-      margin-bottom: 20px;
-    }
-    .login-form .form-control {
-      background-color: #ffffff; /* 밝은 입력 필드 배경색 변경 */
-      color: #000000; /* 어두운 입력 필드 글자색 변경 */
-      border: 1px solid #cccccc; /* 입력 필드 경계선 색상 변경 */
-    }
-    .login-form .btn-primary {
-      background-color: #4e73df; /* 밝은 버튼 색상 유지 */
-      border: none;
-    }
-  </style>
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-</head>
-<body>
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-md-10 content">
-      <div class="login-form">
-        <form th:action="@{/formCsrf}" method="post">
-          <div class="form-group">
-            <label for="username">Username</label>
-            <input type="text" class="form-control" id="username" name="username" required>
-          </div>
-          <div class="form-group">
-            <label for="password">Password</label>
-            <input type="password" class="form-control" id="password" name="password" required>
-          </div>
-          <button type="submit" class="btn btn-primary">Login</button>
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
-</body>
-</html>
