+++ src/main/resources/templates/index.html
@@ -17,6 +17,8 @@
   <button onclick="fetchAndDisplay('/resource/address01')">Manager 페이지 접근</button>
   <button onclick="sendPostRequest()">POST 요청 보내기</button>
 
+  <img style="width: 40px; padding-right: 5px;" src="/images/spring-security-project.png" alt="springSecurity">
+
   <script>
     function fetchAndDisplay(url) {
       fetch(url)
