--- src/main/resources/templates/method.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Button Click Example</title>
-  <script>
-    function sendRequest(url) {
-
-      const dataItems = [
-        { "owner": "user", "isSecure": "false" },
-        { "owner": "admin", "isSecure": "true" },
-        { "owner": "db", "isSecure": "false" }
-      ];
-
-      fetch(url, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(dataItems),
-      })
-              .then(response => {
-                if (response.ok) {
-                  return response.text();
-                }
-                throw new Error('Network response was not ok.');
-              })
-              .then(text => console.log(text))
-              .catch(error => console.error('There was a problem with your fetch operation:', error));
-    }
-  </script>
-</head>
-<body>
-
-<p><button onclick="sendRequest('/writeList')">Pre Account List</button></p>
-<p><button onclick="sendRequest('/writeMap')">Pre Account Map</button></p>
-<p><a href="/readList">Post Account List</a></p>
-<p><a href="/readMap">Post Account Map</a></p>
-
-</body>
-</html>
